feat: add js file for app
