+++ src/js/bigUntestableCode.js
@@ -1,6 +1,7 @@
-import { getItemsRequest, toggleFavoriteRequest } from './requests';
+import { getItemsRequest } from './requests';
 import { showLoader, hideLoader } from './loader';
 import { showError } from './error';
+import { showApp } from './app';
 
 export default () => {
     document.querySelector('#error').style.display = 'none';
@@ -10,26 +11,7 @@ export default () => {
             if (data.result !== 'ok' || typeof data.html === 'undefined') {
                 showError(data);
             } else {
-                const appElement = document.querySelector('#app');
-
-                appElement.innerHTML = data.html;
-                appElement.style.display = 'block';
-                Array.from(appElement.querySelector('button')).forEach(
-                    (button) => {
-                        button.addEventListener('click', (e) => {
-                            e.currentTarget.textContent = 'Загрузка...';
-                            toggleFavoriteRequest(
-                                e.currentTarget.dataset.id,
-                            ).then(({ data: buttonData }) => {
-                                if (buttonData.result === 'set') {
-                                    e.currentTarget.textContent = '🌝';
-                                } else {
-                                    e.currentTarget.textContent = '🌚';
-                                }
-                            });
-                        });
-                    },
-                );
+                showApp(data);
             }
         })
         .catch((e) => {
